fix compilation errors after rebase

+++ Tangerine/Tangerine.UI/PropertyEditors/Vector2PropertyEditor.cs
@@ -31,7 +31,7 @@ namespace Tangerine.UI
 			if (Parser.TryParse(editor.Text, out double newValue)) {
 				DoTransaction(() => {
 					SetProperty<Vector2>((current) => {
-						current[component] = newValue;
+						current[component] = (float)newValue;
 						return current;
 					});
 				});
